rename startplat2omp.h -> starplat2omp.h
