+++ main.cc
@@ -1,19 +1,18 @@
 #include <cstdio>
 #include <cstdlib>
-#include <execinfo.h>
 #include <cstring>
+#include <execinfo.h>
 #include <signal.h>
-
 
 #include "ast/ast.h"
 #include "ast/visitor.h"
+#include "avial.tab.h"
 #include "codegen/astdump.h"
 #include "codegen/starplatIR.h"
-#include "avial.tab.h"
 
-#include "mlir/IR/MLIRContext.h"
 #include "mlir/IR/Builders.h"
 #include "mlir/IR/BuiltinOps.h"
+#include "mlir/IR/MLIRContext.h"
 
 // #include "tblgen2/StarPlatOps.cpp.inc"
 
@@ -23,46 +22,43 @@
 
 #include "starplatDialect/includes/StarPlatDialect.h"
 
+#include "mlir/Conversion/ControlFlowToLLVM/ControlFlowToLLVM.h"
 #include "mlir/Conversion/FuncToLLVM/ConvertFuncToLLVM.h"
 #include "mlir/Conversion/MemRefToLLVM/MemRefToLLVM.h"
 #include "mlir/Conversion/SCFToControlFlow/SCFToControlFlow.h"
-#include "mlir/Conversion/ControlFlowToLLVM/ControlFlowToLLVM.h"
 
-#include "starplatDialect/lowerings/startplat2omp.h"
+#include "starplatDialect/lowerings/starplat2omp.h"
 
 // #include "lowerings/startplat2omp.h"
 
 #define DEBUG_TYPE "dialect-conversion"
 
 extern int yyparse();
-extern FILE *yyin;
-ASTNode *root;
+extern FILE* yyin;
+ASTNode* root;
 
 void signalHandler(int sig) {
-    void *array[10];
+    void* array[10];
     size_t size;
-    
+
     size = backtrace(array, 10);
-    
+
     fprintf(stderr, "Error: signal %d:\n", sig);
     backtrace_symbols_fd(array, size, 0);
     exit(1);
 }
 
-int main(int argc, char *argv[])
-{
+int main(int argc, char* argv[]) {
     root = nullptr;
-    signal(SIGSEGV, signalHandler);  
+    signal(SIGSEGV, signalHandler);
 
-    if (argc < 2)
-    {
+    if (argc < 2) {
         printf("%s usage\n%s <file name>\n", argv[0], argv[0]);
         return 0;
     }
 
-    FILE *file = fopen(argv[1], "r");
-    if (!file)
-    {
+    FILE* file = fopen(argv[1], "r");
+    if (!file) {
         printf("Cannot open file. \n");
         return 0;
     }
@@ -73,29 +69,27 @@ int main(int argc, char *argv[])
 
     printf("Parsing Complete\n");
 
-    CodeGen *codegen = new CodeGen;
+    CodeGen* codegen = new CodeGen;
 
     if (root != nullptr)
         root->Accept(codegen);
 
-    StarPlatCodeGen *starplatcodegen = new StarPlatCodeGen;
+    StarPlatCodeGen* starplatcodegen = new StarPlatCodeGen;
 
     if (root != nullptr)
         root->Accept(starplatcodegen, starplatcodegen->getSymbolTable());
-    else
-    {
+    else {
         llvm::errs() << "Error in DSL Code\n";
         exit(0);
     }
 
-    //starplatcodegen->print();
+    // starplatcodegen->print();
 
     PassManager pm(starplatcodegen->getContext());
     pm.addPass(mlir::starplat::createConvertStartPlatIRToOMPPass());
 
     // // RUN the pass on the module
-    if (mlir::failed(pm.run(starplatcodegen->getModule()->getOperation())))
-    {
+    if (mlir::failed(pm.run(starplatcodegen->getModule()->getOperation()))) {
         llvm::errs() << "StarPlat → OMP lowering failed\n";
         return 0;
     }
